[FIX]- BMP-JPEG conversion
In the following Code there was an issue that the converted image's colour was changed and the original format of the image was refused applying a blue filter on it.

Which was due to the BMP colour pallet being different to the one of JPEG, which requires the use of a converter.  Which was inproper.