[ADD]/ [CHANGE]- Filter call/ Image Type
[Change]
Filter calls are now changed so that the INPUT would be specified before the functionality itself, and the there are changes to the Output so that you can easily change the name within the function and it adds that information to the path by using " %s " which waits for the name to be defined so future thinking that the user could rename the output as well as makes testing easier if you wish to change it within the Specified filter.

[ADD]
Additionally defined an overlay for the INPUT image and output image, which will later on define which file type the image currently is and displays that on the top of the image.
=PLACEHOLDER!

Removed padding that was previously added to the images to center them.